Rename 'discover' command to 'scan' - fixes #69
