+++ cmd/reliably/root.go
@@ -53,8 +53,7 @@ func NewCmdRoot() *cobra.Command {
 		SilenceErrors:     true, // quiet errors down stream
 		SilenceUsage:      true, //silence usage when an error occurs
 		DisableAutoGenTag: true, // disables auto gen tag "Auto generated by spf13/cobra..."
-		Example: heredoc.Doc(`
-			$ reliably discover`),
+		//Example: heredoc.Doc(``),
 		Annotations: map[string]string{
 			"help:feedback": heredoc.Doc(`
 
@@ -89,11 +88,12 @@ Environment variables:
 
 	// Add commands
 	cmd.AddCommand(NewCmdAuth())
-	cmd.AddCommand(NewCmdDiscover())
+	cmd.AddCommand(NewCmdDiscover()) // deprecated
 	cmd.AddCommand(NewCmdCompletion())
 	cmd.AddCommand(NewCmdVersion())
 	cmd.AddCommand(NewCmdWorkflow())
 	cmd.AddCommand(NewCmdHistory())
+	cmd.AddCommand(NewCmdScan())
 
 	//Help topics
 	cmd.AddCommand(NewHelpTopic("environment"))
