+++ core/workflows/circleci.go
@@ -12,7 +12,7 @@ var circleci_Template string = heredoc.Doc(`
 version: 2.1
 
 jobs:
-  discover:
+  scan:
     docker:
       - image: ghcr.io/reliablyhq/cli/cli:latest
         environment:
@@ -20,12 +20,12 @@ jobs:
     working_directory: /home
     steps:
       - checkout # check out the code in the project directory
-      - run: reliably discover .
+      - run: reliably scan .
 
 workflows:
   reliably:
     jobs:
-      - discover
+      - scan
 `)
 
 var circleci_AccessTokenHelp string = `
@@ -46,11 +46,11 @@ func insertReliablyToCircleci(wf workflow, tpl workflow) error {
 	if _, ok := wf["jobs"]; !ok {
 		wf["jobs"] = make(map[interface{}]interface{})
 	}
-	if _, found := wf["jobs"].(map[interface{}]interface{})["discover"]; found {
-		return errors.New("Your workflow already contains a 'discover' job")
+	if _, found := wf["jobs"].(map[interface{}]interface{})["scan"]; found {
+		return errors.New("Your workflow already contains a 'scan' job")
 	}
 	jobs := tpl["jobs"].(map[interface{}]interface{})
-	wf["jobs"].(map[interface{}]interface{})["discover"] = jobs["discover"]
+	wf["jobs"].(map[interface{}]interface{})["scan"] = jobs["scan"]
 
 	if _, ok := wf["workflows"]; !ok {
 		wf["workflows"] = make(map[interface{}]interface{})
